MB Sorting
The expected behavior for both portals is:

Default state - deals are sorted based on latest created application.
Default state is when you login and do not sort anything. Currently
latest created deals on MB do not appear at the top of the table but
somewhere in the middle which is wrong. Also please notice date of
creation.
Ascending order - when you click on column's heading one time.
Descending order - when you click on column's heading second time.
Back to default order - when you click on column's heading third time.